Increments strjoin

+++ ft_strjoin.c
@@ -6,7 +6,7 @@
 /*   By: lusantor <lusantor@student.42.rio>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/26 23:36:23 by lusantor          #+#    #+#             */
-/*   Updated: 2022/06/26 21:17:02 by lusantor         ###   ########.fr       */
+/*   Updated: 2022/06/26 21:18:50 by lusantor         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strjoin(char const *s1, char const *s2)
 {
 	char	*str;
-	int		size;
+	size_t	size;
 
 	size = ft_strlen(s1) + ft_strlen(s2) + 1;
 	str = malloc(size);
